Merged PR 91899: Add COEFFICIENT unit, update descriptions for PERCENTAGE units and MONETARY_U...
Add COEFFICIENT unit, update descriptions for PERCENTAGE units and MONETARY_UNIT, and change display label for MONETARY_UNIT and DECIMAL_PERCENT

I updated COEFFICIENT and DECIMAL_PERCENT to have a display label.  This means you have to set the display label to an empty string in your KOQ or format definition.  I made this change because a unit without a display label will show the invariant name in the UI when the user goes to pick a unit.

Additionally we found a bug in ECObjects that drops the description when the display label is empty.  That is being fixed independently ... this should hopefully be the end of the troubles with this schema release.