Merged PR 123085: Schema Consumption Process
By default, the latest released version of a schema will be consumed and if no released version is available then WIP version will be consumed. If someone needs to consume a WIP version of one or more schema's then set an environment variable WIP_SCHEMAS with a semi-colon delimited list of schema names. To specify all WIP schemas use the character '*'.

This PR also contains fixes for making this process successful on Windows, Linux and Mac.

Related work items: #334318