Merged PR 126565: Schema import tests for multiple schemas
Added a way to run existing iModelSchemaValidation tests to run for sets of multiple schemas together. The Developer can specify these sets through json with an example provided in the code.

The test imports the schemas in the order specified in json. Currently only full schema names in bis-schemas repo's released schemas can be specified in json.

Related work items: #486856