Merged PR 159675: InformationRecordElement-based implementation of the ILinearLocationElement mix-in
Introducing InformationRecordElement-based implementations of the ILinearLocationElement mix-in, including a base class (LinearLocationRecordElement) and concrete class (LinearLocationRecord). The latter enables linear-location of other classes not inherently linearly-located, like the existing LinearLocation class (a bis:SpatialLocationElement subclass). It enables the same functionality when the latter is not appropriate (e.g. the linear-location element does not need to or can't be "geometrical" - subclass of bis:GeometricElement).