Add properties for PileWallPierCapBaseAspect
This class is used as a base class for PileCapPierCapAspect and StemWallPierCapAspect,
which contain properties for caps of type Pile Cap and Stem Wall. But these
two classes should also contain the properties from RegularPierCapBaseAspect
which are displayed in OBM for caps of type Pile Cap and Stem Wall.
Because we cannot change the class hierarchy, we have to duplicate them.